hwyy-frida-dump/requirements.txt
frida-tools~=13.6.1
pycryptodome~=3.22.0
